+++ lib/interval.h
@@ -3,11 +3,11 @@
 
 class interval {
   public:
-    double min, max;
+    float min, max;
 
     interval() : min(+infinity), max(-infinity) {} // Default interval is empty
 
-    interval(double min, double max) : min(min), max(max) {}
+    interval(float min, float max) : min(min), max(max) {}
 
     interval(const interval& a, const interval& b) {
         // Create the interval tightly enclosing the two input intervals.
@@ -15,25 +15,25 @@ class interval {
         max = a.max >= b.max ? a.max : b.max;
     }
 
-    double size() const {
+    float size() const {
         return max - min;
     }
 
-    bool contains(double x) const {
+    bool contains(float x) const {
         return min <= x && x <= max;
     }
 
-    bool surrounds(double x) const {
+    bool surrounds(float x) const {
         return min < x && x < max;
     }
 
-    double clamp(double x) const {
+    float clamp(float x) const {
         if (x < min) return min;
         if (x > max) return max;
         return x;
     }
 
-    interval expand(double delta) const {
+    interval expand(float delta) const {
         auto padding = delta/2;
         return interval(min - padding, max + padding);
     }
