+++ lib/sphere.h
@@ -6,11 +6,11 @@
 class sphere : public hittable {
   public:
     // Stationary Sphere
-    sphere(const point3& static_center, double radius, shared_ptr<material> mat)
+    sphere(const point3& static_center, float radius, shared_ptr<material> mat)
       : center(static_center, vec3(0,0,0)), radius(std::fmax(0,radius)), mat(mat) {}
 
     // Moving Sphere
-    sphere(const point3& center1, const point3& center2, double radius, shared_ptr<material> mat)
+    sphere(const point3& center1, const point3& center2, float radius, shared_ptr<material> mat)
       : center(center1, center2 - center1), radius(std::fmax(0,radius)), mat(mat) {}
     
     bool hit(const ray& r, interval ray_t, hit_record& rec) const override {
@@ -46,7 +46,7 @@ class sphere : public hittable {
     aabb bounding_box() const override { return bbox; }
 
   private:
-    double radius;
+    float radius;
     shared_ptr<material> mat;
     aabb bbox;
     ray center;
