+++ lib/texture.h
@@ -7,16 +7,16 @@ class texture {
   public:
     virtual ~texture() = default;
 
-    virtual color value(double u, double v, const point3& p) const = 0;
+    virtual color value(float u, float v, const point3& p) const = 0;
 };
 
 class solid_color : public texture {
   public:
     solid_color(const color& albedo) : albedo(albedo) {}
 
-    solid_color(double red, double green, double blue) : solid_color(color(red,green,blue)) {}
+    solid_color(float red, float green, float blue) : solid_color(color(red,green,blue)) {}
 
-    color value(double u, double v, const point3& p) const override {
+    color value(float u, float v, const point3& p) const override {
         return albedo;
     }
